requirements.txt
pandas
numpy
seaborn
scikit-learn
catboost
xgboost
dill
Flasks
boto3
gunicorn
# -e .